ml-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
numpy==1.24.3
scikit-learn==1.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
